Changes
The background is now teal. The stroke is a random color each time it opens. The fill is a random value of red. Speed of the circle is slower until the mouse is pressed, than it gets faster.  The circle size is smaller until the mouse is pressed, the ellipse gets wider. Clicking the mouse also changed the tragectory of the ellipse.